+++ src/generators/RelativeDateGenerator.js
@@ -3,10 +3,8 @@ const BaseGenerator = require('./BaseGenerator');
 const { add } = require('date-fns');
 
 class RelativeDateGenerator extends BaseGenerator {
-  generate({ session, params }) {
-    const value = params ?
-      add(session.now, params)
-      : new Date(session.now);
+  generate({ now, params }) {
+    const value = params ? add(now, params) : new Date(now);
     debug('Generated relative date{%o}', value);
     return value;
   }