+++ test/TestDataFactory.test.js
@@ -3,10 +3,8 @@ const { mixed, object, string, number, date } = require('yup');
 
 describe('TestDataFactory', () => {
 
-  let testDataFactory;
-
   beforeEach(() => {
-    testDataFactory = TestDataFactory.init();
+    TestDataFactory.init();
   });
 
   it('should generate test data without options', async () => {
@@ -16,8 +14,8 @@ describe('TestDataFactory', () => {
       otherwise: number().min(2).max(2).integer().required().example(),
     }).example()
 
-    const mandatory = await testDataFactory.generate(schema, { context: { required: true } });
-    const optional = await testDataFactory.generate(schema, { context: { required: false } });
+    const mandatory = await TestDataFactory.generate(schema, { context: { required: true } });
+    const optional = await TestDataFactory.generate(schema, { context: { required: false } });
     expect(mandatory).to.equal(1);
     expect(optional).to.equal(2);
   })
@@ -29,8 +27,8 @@ describe('TestDataFactory', () => {
       otherwise: number().example().notRequired(),
     }).example()
 
-    const mandatory = await testDataFactory.generate(schema, { context: { required: true, abortEarly: true } });
-    const optional = await testDataFactory.generate(schema, { context: { required: false, abortEarly: true } });
+    const mandatory = await TestDataFactory.generate(schema, { context: { required: true, abortEarly: true } });
+    const optional = await TestDataFactory.generate(schema, { context: { required: false, abortEarly: true } });
 
     expect(mandatory).to.be.a('number');
     expect(optional).to.be.null;
@@ -42,7 +40,7 @@ describe('TestDataFactory', () => {
       b: string().matches(/wobble/).required(),
     }).example();
     try {
-      await testDataFactory.generateValid(schema);
+      await TestDataFactory.generateValid(schema);
       expect.fail('Should have reported a validation error')
     } catch (err) {
       expect(err.errors.length).to.equal(1)
@@ -55,61 +53,61 @@ describe('TestDataFactory', () => {
       b: string().matches(/wobble/).required(),
     }).example();
     try {
-      await testDataFactory.generateValid(schema, { abortEarly: false });
+      await TestDataFactory.generateValid(schema, { abortEarly: false });
       expect.fail('Should have reported a validation error')
     } catch (err) {
       expect(err.errors.length).to.equal(2)
     }
   })
 
-  it('should report missing generators', async () => {
+  it('should report missing id generators', async () => {
     const schema = string().example({ generator: 'missing' });
-    await expect(testDataFactory.generateValid(schema))
-      .to.be.rejectedWith('No such generator: \'missing\'');
+    await expect(TestDataFactory.generateValid(schema))
+      .to.be.rejectedWith('Unable to resolve generator from [\'missing\']');
   })
 
-  it('should report unresolvable generators', async () => {
-    testDataFactory.removeGenerator('date');
+  it('should report missing type generators', async () => {
+    TestDataFactory.removeGenerator('date');
     const schema = date().meta({ type: 'missing' }).example();
-    await expect(testDataFactory.generateValid(schema))
+    await expect(TestDataFactory.generateValid(schema))
       .to.be.rejectedWith('Unable to resolve generator from [\'missing\', \'date\']');
   })
 
   it('should select generator by id', async () => {
-    testDataFactory.addGenerator('wibble', CustomGenerator);
+    TestDataFactory.addGenerator('wibble', CustomGenerator);
     const schema = string().example({ generator: 'wibble' });
-    const value = await testDataFactory.generateValid(schema);
+    const value = await TestDataFactory.generateValid(schema);
     expect(value).to.equal('WIBBLE');
   })
 
   it('should select generator by meta.type', async () => {
-    testDataFactory.addGenerator('wibble', CustomGenerator);
+    TestDataFactory.addGenerator('wibble', CustomGenerator);
     const schema = string().meta({ type: 'wibble' }).example();
-    const value = await testDataFactory.generateValid(schema);
+    const value = await TestDataFactory.generateValid(schema);
     expect(value).to.equal('WIBBLE');
   })
 
   it('should select generator by type', async () => {
     const schema = string().example();
-    const value = await testDataFactory.generateValid(schema);
+    const value = await TestDataFactory.generateValid(schema);
     expect(value).to.be.a('string');
   })
 
   it('should tolerate missing generators for meta.type', async () => {
     const schema = string().meta({ type: 'missing' }).example();
-    const value = await testDataFactory.generateValid(schema);
+    const value = await TestDataFactory.generateValid(schema);
     expect(value).to.be.a('string');
   })
 
   it('should report validation errors when using generateValid', async () => {
     const schema = string().matches('does not match').example();
-    await expect(testDataFactory.generateValid(schema))
+    await expect(TestDataFactory.generateValid(schema))
       .to.be.rejectedWith('this must match the following: "does not match"');
   });
 
   it('should permit validation errors when using generate', async () => {
     const schema = string().matches('does not match').example();
-    await expect(testDataFactory.generate(schema))
+    await expect(TestDataFactory.generate(schema))
       .to.not.be.rejected();
   });
 
@@ -124,76 +122,69 @@ describe('TestDataFactory', () => {
     await expect(schema.validate('valid')).to.not.be.rejected()
   });
 
-  it('should provide a harmless noop method', async () => {
-    TestDataFactory.stub();
-    const schema = string().example();
-    const value = schema.cast('WIBBLE');
-    expect(value).to.equal('WIBBLE');
-  });
-
-  it('should notify via example event', async () => {
-    const schema = string().oneOf(['WIBBLE']).example();
-    testDataFactory.session.once('example', event => {
-      expect(event.value).to.equal('WIBBLE');
-      event.value = 'wobble';
-    });
-    const value = await testDataFactory.generate(schema);
-    expect(value).to.equal('wobble');
-  })
-
-  it('should notify via renamed example event', async () => {
-    testDataFactory = TestDataFactory.init({ methodName: 'generate' });
-    const schema = string().oneOf(['WIBBLE']).generate();
-    testDataFactory.session.once('generate', event => {
-      expect(event.value).to.equal('WIBBLE');
-      event.value = 'wobble';
-    });
-    const value = await testDataFactory.generate(schema);
-    expect(value).to.equal('wobble');
-  })
-
   it('should notify via type event', async () => {
     const schema = string().oneOf(['WIBBLE']).example();
-    testDataFactory.session.once('string', event => {
+    TestDataFactory.session.once('string', event => {
       expect(event.value).to.equal('WIBBLE');
       event.value = 'wobble';
     });
-    const value = await testDataFactory.generate(schema);
+    const value = await TestDataFactory.generate(schema);
     expect(value).to.equal('wobble');
   })
 
   it('should notify via meta.type event', async () => {
     const schema = string().meta({ type: 'metatype' }).oneOf(['WIBBLE']).example();
-    testDataFactory.session.once('metatype', event => {
+    TestDataFactory.session.once('metatype', event => {
       expect(event.value).to.equal('WIBBLE');
       event.value = 'wobble';
     });
-    const value = await testDataFactory.generate(schema);
+    const value = await TestDataFactory.generate(schema);
     expect(value).to.equal('wobble');
   })
 
   it('should notify via id event', async () => {
-    testDataFactory.addGenerator('wibble', CustomGenerator);
+    TestDataFactory.addGenerator('wibble', CustomGenerator);
     const schema = string().example({ generator: 'wibble' });
-    testDataFactory.session.once('wibble', event => {
+    TestDataFactory.session.once('wibble', event => {
       expect(event.value).to.equal('WIBBLE');
       event.value = 'wobble';
     });
-    const value = await testDataFactory.generate(schema);
+    const value = await TestDataFactory.generate(schema);
     expect(value).to.equal('wobble');
   })
 
-  it('should reset', async () => {
-    const session = testDataFactory.session;
+  it('should reinitialise', async () => {
+    const session = TestDataFactory.session;
     session.setProperty('wibble', 'wobble');
-    session.on('example', () => {
+    session.on('some-event', () => {
       expect.fail('Event handler was not reset');
     })
 
-    testDataFactory.reset();
+    TestDataFactory.init();
 
-    expect(session.emit('example')).to.equal(false);
+    expect(session.emit('some-event')).to.equal(false);
     expect(session.getProperty('wibble')).to.be.undefined;
+  })
+
+  it('should support adding multiple generators', async () => {
+    TestDataFactory.init().addGenerators({
+      wibble: CustomGenerator
+    });
+    const schema = string().example({ generator: 'wibble' });
+    const value = await TestDataFactory.generate(schema);
+    expect(value).to.equal('WIBBLE');
+  })
+
+  it('should support adding multiple generators during intialisation', async () => {
+    TestDataFactory.init({
+      generators: {
+        wibble: CustomGenerator
+      }
+    });
+
+    const schema = string().example({ generator: 'wibble' });
+    const value = await TestDataFactory.generate(schema);
+    expect(value).to.equal('WIBBLE');
   })
 
   class CustomGenerator {