+++ test/function.test.js
@@ -4,15 +4,13 @@ const TestDataFactory = require('../src/TestDataFactory');
 
 describe('function generator', () => {
 
-  let testDataFactory;
-
   beforeEach(() => {
-    testDataFactory = TestDataFactory.init();
+    TestDataFactory.init();
   });
 
   it('should generate values using the specified function', async () => {
     const schema = number().example({ generator: 'function' }, () => Math.random());
-    const { counts, values } = await sample(1000, () => testDataFactory.generateValid(schema));
+    const { counts, values } = await sample(1000, () => TestDataFactory.generateValid(schema));
 
     expect(counts.length).to.be.above(900);
     expectMostlyFloatingPoints(values);
@@ -20,7 +18,7 @@ describe('function generator', () => {
 
   it('should accept fn as an alias for function', async () => {
     const schema = number().example({ generator: 'fn' }, () => Math.random());
-    const { counts, values } = await sample(1000, () => testDataFactory.generateValid(schema));
+    const { counts, values } = await sample(1000, () => TestDataFactory.generateValid(schema));
 
     expect(counts.length).to.be.above(900);
     expectMostlyFloatingPoints(values);
@@ -28,7 +26,7 @@ describe('function generator', () => {
 
   it('should supply the chance instance', async () => {
     const schema = number().example({ generator: 'fn' }, ({ chance }) => chance.integer());
-    const { counts, values } = await sample(1000, () => testDataFactory.generateValid(schema));
+    const { counts, values } = await sample(1000, () => TestDataFactory.generateValid(schema));
 
     expect(counts.length).to.be.above(900);
     expectAllIntegers(values);