+++ test/literal.test.js
@@ -5,9 +5,9 @@ const TestDataFactory = require('../src/TestDataFactory');
 describe('literal generator', () => {
 
   it('should generate literal values', async () => {
-    const testDataFactory = TestDataFactory.init();
+    TestDataFactory.init();
     const schema = string().example({ generator: 'literal' }, 'wibble');
-    const value = await testDataFactory.generateValid(schema);
+    const value = await TestDataFactory.generateValid(schema);
     expect(value).to.equal('wibble');
   })
 });